更改maniqa结构将texture as query输入模仿dual论文做法

+++ freeGPU.py
@@ -1,7 +1,7 @@
 import os
 
-
 result = os.popen("fuser -v /dev/nvidia*").read()
 results = result.split()
 for pid in results:
-    os.system(f"kill -9 {int(pid)}")
+    if pid.isdigit():  # 检查字符串是否为数字
+        os.system(f"kill -9 {int(pid)}")
